requirements.txt
# Core dependencies for VeilTrader AI
web3==6.15.0
requests==2.31.0
python-dotenv==1.0.0
groq==0.4.1
aiohttp==3.9.0
fastapi==0.109.0
uvicorn==0.27.0
# Optional for Streamlit demo
streamlit==1.32.0
plotly==5.18.0
pandas==2.2.0